requirements.txt
python-dotenv
supervisely

# formatter
black

# sorter
isort
